Enemy Health System
Enemies now have a life system and when they collide with a player's projectile, their life is reduced; when it reaches zero, they self-destruct.

+++ Assets/Game/Scripts/Entity/BowBehaviour.cs
@@ -10,7 +10,7 @@ public class BowBehaviour : MonoBehaviour
     private float _coolDown;
 
     [SerializeField]
-    private float _damage;
+    private int _damage;
 
     [SerializeField]
     private float _projectileSpeed;
